requirements.txt
langgraph
langchain-groq
langchain-community
duckduckgo-search
python-dotenv